+++ src/qiskit_iqm/iqm_provider.py
@@ -14,13 +14,17 @@
 """Qiskit Backend Provider for IQM backends.
 """
 from importlib.metadata import version
+from functools import reduce
 from typing import Optional, Union
+import warnings
 
 from iqm_client import Circuit, Instruction, IQMClient
+from iqm_client.util import to_json_dict
 import numpy as np
 from qiskit import QuantumCircuit
-from qiskit.providers import Options
+from qiskit.providers import JobStatus, JobV1, Options
 
+from qiskit_iqm.fake_backends import IQMFakeAdonis
 from qiskit_iqm.iqm_backend import IQMBackendBase
 from qiskit_iqm.iqm_job import IQMJob
 from qiskit_iqm.qiskit_to_iqm import MeasurementKey
@@ -59,7 +63,10 @@ class IQMBackend(IQMBackendBase):
         calibration_set_id = options.get('calibration_set_id', self.options.calibration_set_id)
 
         circuits_serialized: list[Circuit] = [self.serialize_circuit(circuit) for circuit in circuits]
-        qubit_mapping = {str(idx): qb for idx, qb in self._idx_to_qb.items()}
+        used_indices: set[int] = reduce(
+            lambda qubits, circuit: qubits.union(set(int(q) for q in circuit.all_qubits())), circuits_serialized, set()
+        )
+        qubit_mapping = {str(idx): qb for idx, qb in self._idx_to_qb.items() if idx in used_indices}
         uuid = self.client.submit_circuits(
             circuits_serialized, qubit_mapping=qubit_mapping, calibration_set_id=calibration_set_id, shots=shots
         )
@@ -121,7 +128,62 @@ class IQMBackend(IQMBackendBase):
                     f'You need to transpile the circuit before execution.'
                 )
 
-        return Circuit(name=circuit.name, instructions=instructions, metadata=circuit.metadata)
+        try:
+            metadata = to_json_dict(circuit.metadata)
+        except ValueError:
+            warnings.warn(
+                f'Metadata of circuit {circuit.name} was dropped because it could not be serialised to JSON.',
+            )
+            metadata = None
+
+        return Circuit(name=circuit.name, instructions=instructions, metadata=metadata)
+
+
+class IQMFacadeBackend(IQMBackend):
+    """Facade backend for mimicking the execution of quantum circuits on IQM quantum computers. Allows to submit a
+     circuit to the IQM server, and if the execution was successful, performs a simulation with a respective IQM noise
+     model locally, then returns the simulated results.
+
+    Args:
+        client: client instance for communicating with an IQM server
+        **kwargs: optional arguments to be passed to the parent Backend initializer
+    """
+
+    def __init__(self, client: IQMClient, **kwargs):
+        self.fake_adonis = IQMFakeAdonis()
+        target_architecture = client.get_quantum_architecture()
+
+        if not self.fake_adonis.validate_compatible_architecture(target_architecture):
+            raise ValueError('Quantum architecture of the remote quantum computer does not match Adonis.')
+
+        super().__init__(client, **kwargs)
+        self.client = client
+
+    def _validate_no_empty_cregs(self, circuit):
+        """Returns True if given circuit has no empty (unused) classical registers, False otherwise."""
+        cregs_utilization = dict.fromkeys(circuit.cregs, 0)
+        used_cregs = [circuit.find_bit(i.clbits[0]).registers[0][0] for i in circuit.data if len(i.clbits) > 0]
+        for creg in used_cregs:
+            cregs_utilization[creg] += 1
+
+        if 0 in cregs_utilization.values():
+            return False
+        return True
+
+    def run(self, run_input: Union[QuantumCircuit, list[QuantumCircuit]], **options) -> JobV1:
+        circuits = [run_input] if isinstance(run_input, QuantumCircuit) else run_input
+        circuits_validated_cregs: list[bool] = [self._validate_no_empty_cregs(circuit) for circuit in circuits]
+        if not all(circuits_validated_cregs):
+            raise ValueError(
+                'One or more circuits contain unused classical registers. This is not allowed for Facade simulation, '
+                'see user guide.'
+            )
+
+        iqm_backend_job = super().run(run_input, **options)
+        iqm_backend_job.result()  # get and discard results
+        if iqm_backend_job.status() == JobStatus.ERROR:
+            raise RuntimeError('Remote execution did not succeed.')
+        return self.fake_adonis.run(run_input, **options)
 
 
 class IQMProvider:
@@ -143,7 +205,13 @@ class IQMProvider:
         self.url = url
         self.user_auth_args = user_auth_args
 
-    def get_backend(self) -> IQMBackend:
-        """An IQMBackend instance associated with this provider."""
+    def get_backend(self, name=None) -> Union[IQMBackend, IQMFacadeBackend]:
+        """An IQMBackend instance associated with this provider.
+
+        Args:
+            name: optional name of a custom facade backend
+        """
         client = IQMClient(self.url, client_signature=f'qiskit-iqm {version("qiskit-iqm")}', **self.user_auth_args)
+        if name == 'facade_adonis':
+            return IQMFacadeBackend(client)
         return IQMBackend(client)